add machine arg to setup script

+++ setup.sh
@@ -3,15 +3,21 @@ set -xeuo pipefail
 pushd $(dirname $0)
 SCRIPTPATH=$(pwd)
 
-git submodule update --init --recursive --depth 1 || true
+: ${1?"Usage: $0 machine-name"}
+test -e $SCRIPTPATH/nix/machines/${1}.nix || exit 99
 
 sudo rm -rf /etc/nixos/configuration.nix
 sudo ln -s $SCRIPTPATH/nix/configuration.nix /etc/nixos/configuration.nix
+
+rm -rf $SCRIPTPATH/nix/machine.nix
+ln -s $SCRIPTPATH/nix/machines/${1}.nix $SCRIPTPATH/nix/machine.nix
 
 sudo find /etc/nixos -mindepth 1 -not -name 'hardware-configuration.nix' -delete
 sudo ln -s $SCRIPTPATH/nix/* /etc/nixos
 
 sudo nixos-rebuild switch --upgrade
+
+git submodule update --init --recursive --depth 1 || true
 
 rm -rf $HOME/.dotfiles
 ln -s $SCRIPTPATH/. $HOME/.dotfiles
